requirements.txt
transformers==4.46.3

torchvision
Pillow
numpy
requests
addict
tqdm
tokenizers==0.20.3
pymupdf
img2pdf
einops
easydict
matplotlib
